[refactor]: valueObjectの命名をパスカルケースに変更
