+++ apps/backend/src/domains/repositories/ChatRoomRepository.ts
@@ -1,6 +1,6 @@
 import { ChatRoom } from '../entities/ChatRoom';
-import { UserId } from '../valueObject/userId';
-import { ChatRoomName } from '../valueObject/chatRoomName';
+import { UserId } from '../valueObject/UserId';
+import { ChatRoomName } from '../valueObject/ChatRoomName';
 
 export interface ChatRoomRepository {
   findMany(userId: UserId): Promise<ChatRoom[]>;
